requirements.txt
Flask
requests
pymongo
firebase-admin
python-dotenv
yfinance
numpy
pandas
scikit-learn
tensorflow
gunicorn
transformers
torch